GUI detection 모듈화
