+++ py_model/detection.py
@@ -1,0 +1,2 @@
+def detect(roi_image):
+    return 0
